hefty refactor of file structure to better represent how things flow
