+++ src/deploy/helpers/flatten.js
