+++ src/deploy/lambda-one/00-validate.js
