+++ src/deploy/lambda-one/03-copy-shared.js
